Errors fix
