+++ front/src/components/Header/PropTypes/userName.js
@@ -1,0 +1,8 @@
+import PropTypes from 'prop-types';
+
+const objectUsername = PropTypes.shape({
+    firstName: PropTypes.string.isRequired,
+    secondName: PropTypes.string.isRequired
+});
+
+export default objectUsername;
