+++ front/src/components/Header/components/AddArticleButton/AddArticleButton.jsx
@@ -1,10 +1,11 @@
 import React from 'react';
 import './AddArticleButton.css';
+import CreateArticleContainer from '../../../UserContainers/CreateArticle/CreateArticle';
 import PropTypes from 'prop-types';
 
 function AddArticleButton({clickAddArticleButton}) {
     return (
-        <button onClick={clickAddArticleButton} className="add_article__button">Add article</button>
+        <button onClick={() => clickAddArticleButton(<CreateArticleContainer/>)} className="add_article__button">Add article</button>
     );
 }
