+++ front/src/components/Header/components/Logo/Logo.jsx
@@ -1,10 +1,11 @@
 import React from 'react';
 import './Logo.css';
 import PropTypes from 'prop-types';
+import ArticlesContainer from '../../../UserContainers/ArticlesPage/ArticlesPage';
 
 function Logo({clickLogo}) {
     return (
-        <h1 onClick={clickLogo} className="logo-title">Swiftchat</h1>
+        <h1 onClick={() => clickLogo(<ArticlesContainer/>)} className="logo-title">Swiftchat</h1>
     );
 }
