+++ front/src/components/Header/components/UserInfo/components/Menu/Menu.jsx
@@ -3,18 +3,19 @@ import './Menu.css';
 import MenuItem from './components/MenuItem';
 import PropTypes from 'prop-types';
 
-function Menu({clickProfile}) {
+function Menu({clickProfile, setUsername}) {
     
     return(
         <ul className="menu">
-            <MenuItem clickMenuItem={clickProfile} text={'Profile'} link={'#'}></MenuItem>
-            <MenuItem text={'LogOut'}></MenuItem>
+            <MenuItem clickMenuItem={clickProfile} text={'Profile'} setUsername={setUsername}/>
+            <MenuItem text={'LogOut'}/>
         </ul>
     );
 }
 
 Menu.propTypes = {
     clickProfile: PropTypes.func.isRequired,
+    setUsername: PropTypes.func.isRequired,
 }
 
 export default Menu;
