+++ front/src/components/Header/components/UserInfo/components/Menu/components/MenuItem.jsx
@@ -1,10 +1,11 @@
 import React from 'react';
 import './MenuItem.css';
 import PropTypes from 'prop-types';
+import MyProfileContainer from '../../../../../../UserContainers/MyProfile/MyProfile';
 
-function MenuItem( {text, clickMenuItem} ) {
+function MenuItem( {text, clickMenuItem, setUsername} ) {
     return(
-        <li className="menu__item" onClick={clickMenuItem}>
+        <li className="menu__item" onClick={() => clickMenuItem(<MyProfileContainer setUsername={setUsername}/>)}>
             {text}
         </li>
     );
@@ -13,6 +14,7 @@ function MenuItem( {text, clickMenuItem} ) {
 MenuItem.propTypes = {
     text: PropTypes.string.isRequired,
     clickMenuItem: PropTypes.func.isRequired,
+    setUsername: PropTypes.func.isRequired,
 }
 
 export default MenuItem;
