+++ front/src/components/UserContainers/ArticlesPage/ArticlesPage.jsx
@@ -5,10 +5,10 @@ import Article from '../../Article/Article';
 function Articles(){
     return(
         <div className="articles-block">
-            <Article></Article>
-            <Article></Article>
-            <Article></Article>
-            <Article></Article>
+            <Article/>
+            <Article/>
+            <Article/>
+            <Article/>
         </div>
     );
 }
